renamed project folders and other classes to meaningful name
