+++ src/test/java/com/example/demoAPI/SmokeTest.java
@@ -1,4 +1,4 @@
-package com.example.oopdemo;
+package com.example.demoAPI;
 import static org.assertj.core.api.Assertions.assertThat;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
